Fix v

+++ v/adler32.v
@@ -1,7 +1,7 @@
 fn main() {
-    mod_adler := 65521
-    mut a := 1
-    mut b := 0
+    mod_adler := u32(65521)
+    mut a := u32(1)
+    mut b := u32(0)
     for i := u32(0); i < 250000000; i++ {
         a = (a + i) % mod_adler
         b = (b + a) % mod_adler
